[Refactor] VideoCreator 제거
